fix: move config to module init
