--- mDeepFRI/config/__init__.py
@@ -1,3 +1,0 @@
-import multiprocessing
-
-CPU_COUNT = multiprocessing.cpu_count()
