--- mDeepFRI/config/names.py
@@ -1,9 +1,0 @@
-# folder names
-TARGET_MMSEQS_DB_NAME = "targetDB"
-SEQUENCES = "seq"
-ATOMS = "atom"
-SEQ_ATOMS_DATASET_PATH = "seq_atom_db"
-
-ALIGNMENTS = "alignments.json"
-MERGED_SEQUENCES = 'merged_sequences.faa'
-MMSEQS_SEARCH_RESULTS = 'mmseqs2_search_results.m8'
