+++ mDeepFRI/pipeline.py
@@ -6,8 +6,7 @@ from typing import List, Tuple
 
 from pysam.libcfaidx import FastxFile
 
-from mDeepFRI.config.names import (ATOMS, SEQ_ATOMS_DATASET_PATH,
-                                   TARGET_MMSEQS_DB_NAME)
+from mDeepFRI import ATOMS, SEQ_ATOMS_DATASET_PATH, TARGET_MMSEQS_DB_NAME
 from mDeepFRI.CPP_lib import libAtomDistanceIO  # type: ignore[attr-defined]
 from mDeepFRI.predict import Predictor
 from mDeepFRI.utils.fasta_file_io import SeqFileLoader
