+++ mDeepFRI/utils/fasta_file_io.py
@@ -2,7 +2,7 @@ import pathlib
 
 from pysam.libcfaidx import FastxFile
 
-from mDeepFRI.config.names import ATOMS, SEQUENCES
+from mDeepFRI import ATOMS, SEQUENCES
 
 
 class SeqFileLoader:
